requirements.txt
pillow
pynput